Add some tests on the path simplification.
